moved temperature tests to h-transport-problem tests
